requirements.txt
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openpyxl==3.1.5
pillow==11.2.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
reportlab==4.4.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
flasgger==0.9.7.1
alembic==1.13.1
